refactor(baro-rs): decompose driver page into per-section subcomponents
The 439-line DriverPage becomes a 68-line orchestrator plus six
focused subcomponents and a rust-highlight module:

  rust-highlight.ts     — pure token scanner, no Vue
  DriverHero.vue        — breadcrumb + proto-chip + name + lede
  DriverSpec.vue        — datasheet readout (i2c, protocol, measurements)
  DriverFeatures.vue    — numbered capability list
  DriverCode.vue        — editor with highlighted Rust example
  DriverIntegration.vue — integration prose + GitHub link
  DriverFleet.vue       — other-driver list with --odc accent

Per-driver `--dc` color stays on the .drv root via computed themeStyle.
chip-pulse keyframe stays in DriverHero where the dot lives. Rust
syntax token classes (.c-comment/.c-string/etc.) are :deep()-scoped
inside DriverCode so the highlightRust() output still gets coloured.

Each subsection owns its own padding explicitly instead of inheriting
from a shared :deep(.act) rule — cross-component scope specificity
was causing act-hero's 5rem padding to lose to the parent's 4rem
shorthand. Keeping padding local avoids the cascade fight.

No visual change: 'bun run test:visual -- --grep baro-rs' green on all
three viewports for both home and driver routes.
